% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpodclustr-package.r
\name{initialImpute}
\alias{initialImpute}
\title{Function for initial imputation for k-means}
\usage{
initialImpute(X)
}
\arguments{
\item{X}{Data matrix containing missing entries whose rows are observations and columns are features}
}
\value{
A data matrix containing no missing entries
}
\description{
\code{initialImpute} Initial imputation for k-means
}
\examples{
p <- 2
n <- 100
k <- 3
sigma <- 0.25
missing <- 0.05
Data <- makeData(p,n,k,sigma,missing)
X <- Data$Missing
X_copy <- initialImpute(X)

}
\author{
Jocelyn T. Chi
}
