% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kmpp}
\alias{kmpp}
\title{k-means++}
\usage{
kmpp(X, k)
}
\arguments{
\item{X}{Data matrix whose rows are observations and columns are features}

\item{k}{Number of clusters.}
}
\value{
A data matrix whose rows contain initial centroids for the k clusters
}
\description{
\code{kmpp} Computes initial centroids via kmeans++
}
\examples{
n <- 10
p <- 2
X <- matrix(rnorm(n*p),n,p)
k <- 3
kmpp(X,k)
}

