% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpodclustr-package.r
\name{makeData}
\alias{makeData}
\title{Make test data}
\usage{
makeData(p, n, k, sigma, missing, seed = 12345)
}
\arguments{
\item{p}{Number of features (or variables)}

\item{n}{Number of observations}

\item{k}{Number of clusters}

\item{sigma}{Variance}

\item{missing}{Desired missingness percentage}

\item{seed}{(Optional) Seed (default seed is 12345)}
}
\description{
\code{makeData} Function for making test data
}
\examples{
p <- 2
n <- 100
k <- 3
sigma <- 0.25
missing <- 0.05

X <- makeData(p,n,k,sigma,missing)$Orig

}
\author{
Jocelyn T. Chi
}
